% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiCMIT.R
\name{EpiCMIT}
\alias{EpiCMIT}
\title{Estimate EpiCMIT scores}
\usage{
EpiCMIT(data.m)
}
\arguments{
\item{data.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}
}
\value{
A list containing the following entries.
\itemize{
\item hyperSC: EpiCMIT-hyper score of each sample.
\item hypoSC: EpiCMIT-hypo score of each sample.
}
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will return EpiCMIT scores.
}
\details{
EpiCMIT is based on two groups of CpGs, a 184 age-associated hypermethylated CpG list and a 1164 hypomethylated CpG list (Duran-Ferrer et al. 2020) . The EpiCMIT function calculates the average DNAm of the two groups separately and returns the EpiCMIT_hyper and EpiCMIT_hypo scores.
}
\examples{
data("LungInv")
epicmit.o<-EpiCMIT(data.m = bmiq.m)
}
\references{
Duran-Ferrer M, Clot G, Nadeu F, et al.
The proliferative history shapes the DNA methylome of B-cell tumors and predicts clinical outcome.
\emph{Nat Cancer} 2020
}
\seealso{
\code{\link[=EpiMitClocks]{EpiMitClocks()}}
}
