% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiTOC1.R
\name{epiTOC1}
\alias{epiTOC1}
\title{Estimate epiTOC1 score}
\usage{
epiTOC1(data.m)
}
\arguments{
\item{data.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}
}
\value{
The epiTOC1 score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will return the epiTOC1 score.
}
\details{
The epiTOC1 score is calculated as the average beta value of 385 CpGs described in Yang et al. 2014. These CpGs are promoter CpG sites that localize to Polycomb group target genes that are unmethylated in 11 different fetal tissue types.
}
\examples{
data("LungInv")
epitoc1.v<-epiTOC1(data.m = bmiq.m)

}
\references{
Yang Z, Wong A, Kuh D, et al.
Correlation of an epigenetic mitotic clock with cancer risk.
\emph{Genome Biol.} 2016
}
\seealso{
\code{\link[=EpiMitClocks]{EpiMitClocks()}}
}
