% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiTOC2.R
\name{epiTOC2}
\alias{epiTOC2}
\title{Estimate epiTOC2 scores}
\usage{
epiTOC2(data.m, ages.v = NULL)
}
\arguments{
\item{data.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}

\item{ages.v}{Optional argument representing the chronological ages or surrogates thereof of the samples. Vector must be of same length as the number of samples.}
}
\value{
A list containing the following entries
\itemize{
\item tnsc: The estimated cumulative number of stem-cell divisions per stem-cell per year and per sample using the full epiTOC2 model.
\item tnsc2: The estimated cumulative number of stem-cell divisions per stem-cell per year and per sample using an approximation of epiTOC2 which assumes all epiTOC2 CpGs have beta-values exactly 0 in the fetal stage.
\item irS: This is returned only if the ages are provided, and gives the estimated average lifetime intrinsic rate of stem-cell division per sample, as derived from epiTOC2
\item irS2: As irS, but for the approximation.
\item irT: The median estimate over all irS values, yielding a median estimate for the intrinsic rate of stem-cell division for the tissue.
\item irT2: As irT, but for the approximation.
}
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and an age vector (optional) and will return the epiTOC2 scores.
}
\details{
Building upon a dynamic model of DNA methylation gain in unmethylated CpG-rich regions, epiTOC2 can directly estimate the cumulative number of stem cell divisions in a tissue. The details of the algorithm are described in Teschendorff et al. 2020.
}
\examples{
data("LungInv")
epitoc2.o<-epiTOC2(data.m = bmiq.m,ages.v = df$Age)

}
\references{
Teschendorff AE.
A comparison of epigenetic mitotic-like clocks for cancer risk prediction.
\emph{Genome Med.} 2020
}
\seealso{
\code{\link[=EpiMitClocks]{EpiMitClocks()}}
}
