% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypoClock.R
\name{HypoClock}
\alias{HypoClock}
\title{Estimate HypoClock score}
\usage{
HypoClock(data.m)
}
\arguments{
\item{data.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}
}
\value{
The HypoClock score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will return the HypoClock score.
}
\details{
The HypoClock is based on hypomethylation at solo-WCGW sites. The score is calculated as 1-meanbeta(678 solo-WCGWs).
}
\examples{
data("LungInv")
hypoclock.v<-HypoClock(data.m = bmiq.m)

}
\references{
Teschendorff AE.
A comparison of epigenetic mitotic-like clocks for cancer risk prediction.
\emph{Genome Med.} 2020
}
\seealso{
\code{\link[=EpiMitClocks]{EpiMitClocks()}}
}
