% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepliTali.R
\name{RepliTali}
\alias{RepliTali}
\title{Estimate RepliTali score}
\usage{
RepliTali(data.m)
}
\arguments{
\item{data.m}{DNAm beta value matrix with rows labeling Illumina 450k/EPIC CpGs and columns labeling samples.}
}
\value{
The RepliTali score of each sample.
}
\description{
This function takes as input an Illumina 450k/EPIC DNAm beta matrix and will return the RepliTali score.
}
\details{
The RepliTali model is described in Endicott et al. 2022. It is based on 87 population doubling associated hypomethylated CpGs.
}
\examples{
data("LungInv")
replitali.v<-RepliTali(data.m = bmiq.m)
}
\references{
Endicott JL, Nolte PA, Shen H, Laird PW.
Cell division drives DNA methylation loss in late-replicating domains in primary human cells.
\emph{Nat Commun.} 2022
}
\seealso{
\code{\link[=EpiMitClocks]{EpiMitClocks()}}
}
